requirements.txt
Flask==1.1.1
python-telegram-bot==12.1.0
python-chess==0.28.3
google-cloud-ndb==0.1.0
google-cloud-logging==1.8.0
CairoSVG==2.4.2
emoji==0.5.4
humanfriendly==4.18